rename plotting to vis
