+++ pipit/vis/core.py
