+++ pipit/vis/util.py
